Split tool: Allow CLI and Python API access
