+++ kissim/comparison/__init__.py
@@ -6,3 +6,4 @@ from .feature_distances import FeatureDistances
 from .feature_distances_generator import FeatureDistancesGenerator
 from .fingerprint_distance import FingerprintDistance
 from .fingerprint_distance_generator import FingerprintDistanceGenerator
+from .kissim_to_newick import kissim_to_newick
